Refactor configmodels, split file per component, fix godocs
